requirements.txt
Django==2.0b1
mysqlclient==1.3.10
wheel==0.30.0
gunicorn==19.7.1
psycopg2==2.7.3.2
django-tinymce==2.6.0